Switch from int to long

+++ src/de/flozo/Primes.java
@@ -1,11 +1,13 @@
 package de.flozo;
 
+import java.util.function.Function;
+
 public class Primes {
 
-    public static void printList(int limit) {
-        int count = 0;
+    public static void printList(long limit) {
+        long count = 0;
         // Check all numbers
-        for (int i = 0; i <= limit; i++) {
+        for (long i = 0; i <= limit; i++) {
             if (Number.isPrimeSqrtOdd(i)) {
                 count++;
                 System.out.println(i);
@@ -15,11 +17,11 @@ public class Primes {
         System.out.println(count + " prime numbers found.");
     }
 
-    public static void printListOdd(int limit) {
+    public static void printListOdd(long limit) {
         System.out.println("2");
-        int count = 1;
+        long count = 1;
         // check odd numbers only
-        for (int i = 3; i <= limit; i += 2) {
+        for (long i = 3; i <= limit; i += 2) {
             if (Number.isPrimeSqrtOdd(i)) {
                 count++;
                 System.out.println(i);
@@ -29,8 +31,8 @@ public class Primes {
         System.out.println(count + " prime numbers found.");
     }
 
-    public static void printList6k(int limit) {
-        int count = 0;
+    public static void printList6k(long limit) {
+        long count = 0;
         if (limit >= 2) {
             count++;
             System.out.println("2");
@@ -39,12 +41,12 @@ public class Primes {
             count++;
             System.out.println("3");
         }
-        int methodLimit = (limit + 1) / 6 + 1;
+        long methodLimit = (limit + 1) / 6 + 1;
         // check numbers of form 6*k+-1 only
-        for (int i = 1; i < methodLimit; i++) {
+        for (long i = 1; i < methodLimit; i++) {
             // define number pair
-            int class1 = 6 * i - 1;
-            int class2 = 6 * i + 1;
+            long class1 = 6 * i - 1;
+            long class2 = 6 * i + 1;
             // check first number
             if (Number.isPrimeSqrtOdd(class1)) {
                 count++;
@@ -59,5 +61,38 @@ public class Primes {
         System.out.println("*****");
         System.out.println(count + " prime numbers found.");
     }
+
+
+//    public static void printListFK(int limit) {
+//        int count = 0;
+//        if (limit >= 2) {
+//            count++;
+//            System.out.println("2");
+//        }
+//        if (limit >= 3) {
+//            count++;
+//            System.out.println("3");
+//        }
+//        int methodLimit = (limit + 1) / 6 + 1;
+//        // check numbers of form 6*k+-1 only
+//        for (int i = 1; i < methodLimit; i++) {
+//            // define number pair
+//            int class1 = 6 * i - 1;
+//            int class2 = 6 * i + 1;
+//            // check first number
+//            if (Number.isPrimeSqrtOdd(class1)) {
+//                count++;
+//                System.out.println(class1);
+//            }
+//            // check second number if <= limit
+//            if ((class2 <= limit) && (Number.isPrimeSqrtOdd(class2))) {
+//                count++;
+//                System.out.println(class2);
+//            }
+//        }
+//        System.out.println("*****");
+//        System.out.println(count + " prime numbers found.");
+//    }
+
 
 }
